feat: PostSortableImage 컴포넌트 분리
